// src/api/generated/vegnett/types.gen.ts
export type TypeVeg = 'Enkel bilveg' | 'Kanalisert veg' | 'Rampe' | 'Rundkjøring' | 'Bilferje' | 'Gang- og sykkelveg' | 'Sykkelveg' | 'Gangveg' | 'Gågate' | 'Fortau' | 'Trapp' | 'Gangfelt' | 'Gatetun' | 'Passasjerferje' | 'Traktorveg' | 'Sti' | 'Annet';

export type Adresse = {
    id: number;
    navn: string;
    adressekode: number;
    sideveg?: boolean;
};

export type AdskilteLop = 'Med' | 'Mot' | 'Nei';

export type Detaljniva = 'Vegtrase' | 'Kjørebane' | 'Kjørefelt' | 'Vegtrase og kjørebane';

export type GeometriMedKvalitet = {
    wkt: string;
    srid: number;
    kvalitet: Kvalitet;
    datafangstdato: string;
    oppdateringsdato?: string;
    verifiseringsdato?: string;
    kommune: number;
    temakode: number;
    lengde: number;
    medium?: SosiMedium;
    medium_nvdb?: number;
};

export type Kontraktsomrade = {
    id: number;
    nummer?: number;
    navn: string;
};

export type Kryssystem = {
    kryssystem: number;
    kryssdel: number;
    retning: Retning;
    trafikantgruppe: Trafikantgruppe;
    meter?: number;
    fra_meter?: number;
    til_meter?: number;
};

export type Kvalitet = {
    datafangstmetode?: string;
    nøyaktighet?: number;
    datafangstmetodeHøyde?: string;
    nøyaktighetHøyde?: number;
    maksimaltAvvik?: number;
    synbarhet?: number;
};

export type Linje = Omit<Segment, 'type'> & {
    veglenkesekvensid: number;
    startposisjon: number;
    sluttposisjon: number;
    kortform: string;
    type: 'Linje';
};

export type Malemetode = 'Metrert' | 'Geometrisk';

export type Metadata = {
    startdato: string;
    sluttdato?: string;
};

export type Punkt = Omit<Segment, 'type'> & {
    veglenkesekvensid: number;
    relativPosisjon: number;
    kortform: string;
    type: 'Punkt';
};

export type Retning = 'MED' | 'MOT';

export type Riksvegrute = {
    enumid: number;
    riksvegrute: string;
};
// ...
export type Sideanlegg = {
    sideanlegg: number;
    sideanleggsdel: number;
    retning: Retning;
    trafikantgruppe: Trafikantgruppe;
    meter?: number;
    fra_meter?: number;
    til_meter?: number;
};

export type Sideposisjon = 'MH' | 'MV' | 'VT' | 'M' | 'H' | 'HT' | 'HV' | 'K' | 'V' | 'L' | 'R0';

export type SosiMedium = 'T' | 'B' | 'L' | 'U' | 'S' | 'O' | 'V' | 'D' | 'I' | 'W' | 'J' | 'X' | 'IKKE_REGISTRERT';

export type Strekning = {
    strekning: number;
    delstrekning: number;
    arm: boolean;
    adskilte_løp: AdskilteLop;
    trafikantgruppe: Trafikantgruppe;
    retning: Retning;
    meter?: number;
    fra_meter?: number;
    til_meter?: number;
    adskilte_løp_nummer?: string;
};

export type Superstedfesting = {
    veglenkesekvensid: number;
    startposisjon: number;
    sluttposisjon: number;
    retning: Retning;
    kjørefelt: Array<string>;
    kjørefeltMetrering?: Array<string>;
    sideposisjon?: Sideposisjon;
};

export type TypeVegSosi = 'kanalisertVeg' | 'enkelBilveg' | 'rampe' | 'rundkjøring' | 'bilferje' | 'passasjerferje' | 'gangOgSykkelveg' | 'sykkelveg' | 'gangveg' | 'gågate' | 'fortau' | 'trapp' | 'gangfelt' | 'gatetun' | 'traktorveg' | 'sti' | 'annet';

export type Vegfase = 'P' | 'A' | 'V' | 'F';

export type Vegkategori = 'E' | 'F' | 'K' | 'P' | 'R' | 'S' | 'U';

export type VeglenkeType = 'Ukjent' | 'DETALJERT' | 'KONNEKTERING' | 'DETALJERT_KONNEKTERING' | 'HOVED';

export type Vegsystem = {
    vegkategori: Vegkategori;
    fase: Vegfase;
    nummer?: number;
};

export type Vegsystemreferanse = {
    vegsystem?: Vegsystem;
    strekning?: Strekning;
    kryssystem?: Kryssystem;
    sideanlegg?: Sideanlegg;
    kortform?: string;
};

export type TopologinivaParameter = 'kjørebane' | 'kjørefelt' | 'vegtrase';

export type NesteSide = {
    start: string;
    href: string;
};

export type SideMetadata = {
    antall?: number;
    returnert: number;
    sidestørrelse: number;
    neste?: NesteSide;
};

export type Topologiniva = 'Vegtrase' | 'Kjørebane' | 'Kjørefelt';
// ...
export type Vegforvalter = {
    enumid: number;
    vegforvalter: string;
};

export type Veglenkesegment = {
    veglenkesekvensid: number;
    /**
     * URL-en til den segmenterte veglenkesekvensen som dette segmentet tilhører
     */
    href: string;
    metadata: Metadata;
    startposisjon: number;
    sluttposisjon: number;
    kortform: string;
    veglenkenummer: number;
    segmentnummer: number;
    /**
     * I V4 så er fiktive noder representert som posisjon@veglenkesekvens
     */
    startnode: string;
    /**
     * I V4 så er fiktive noder representert som posisjon@veglenkesekvens
     */
    sluttnode: string;
    /**
     * Kombinasjon av veglenkesekvensid, veglenkenummer og segmentnummer. Unik for hvert segment.
     */
    referanse: string;
    type: VeglenkeType;
    detaljnivå: Detaljniva;
    typeVeg: TypeVeg;
    typeVeg_sosi: TypeVegSosi;
    målemetode: Malemetode;
    måledato?: string;
    superstedfesting?: Superstedfesting;
    feltoversikt?: Array<string>;
    feltoversiktMetreringsretning?: Array<string>;
    geometri: GeometriMedKvalitet;
    lengde: number;
    fylke: number;
    kommune: number;
    vegsystemreferanse: Vegsystemreferanse;
    kontraktsområder?: Array<Kontraktsomrade>;
    riksvegruter?: Array<Riksvegrute>;
    vegforvaltere?: Array<Vegforvalter>;
    adresse?: Adresse;
    readonly topologinivå: Topologiniva;
};

export type VeglenkesegmenterSide = {
    objekter: Array<Veglenkesegment>;
    metadata: SideMetadata;
};